ingestion/requirements.txt
s3fs
python-dotenv
pandas
sqlalchemy
psycopg2-binary